digikam from trunk : fix crash under FreeBsd
CCBUGS: 131301

svn path=/trunk/extragear/graphics/digikam/; revision=566463

+++ digikam/albumlister.cpp
@@ -323,7 +323,7 @@ void AlbumLister::slotData(KIO::Job*, const QByteArray& data)
     int     albumID;
     QString name;
     QString date;
-    off_t   size;
+    size_t  size;
     QSize   dims;
 
     ImageInfoList newItemsList;
